# apidefinitionsiot.yaml
swagger: "2.0"
info:
  title: "IoT Operations Partner API"
...
  version: '3.1'
  contact:
    name: "Aruba iotops team"
    email: "wei.huang@hpe.com"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /api/v3/zigbee/devices:
    get:
      summary: "Get Zigbee device list"
      description: "get all Zigbee devices\nRequired scope: context; permission: readonly"
      operationId: "DeviceContextService_GetZigbeeDeviceList"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/v3GetZigbeeDeviceListResponse"
        default:
          description: "An unexpected error response."
          schema:
            $ref: "#/definitions/rpcStatus"
      tags:
      - "Zigbee"
...
definitions:
...
  protobufAny:
    type: "object"
    properties:
      '@type':
        type: "string"
    additionalProperties: {}
...
  rpcStatus:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      details:
        type: "array"
        items:
          $ref: "#/definitions/protobufAny"
...
  v3GetZigbeeDeviceListResponse:
    type: "object"
    properties:
      hwId:
        type: "array"
        items:
          type: "string"
...
securityDefinitions:
  ApiKeyAuth:
    type: "apiKey"
    name: "apikey"
    in: "header"
security:
- ApiKeyAuth: []